Non-matching decomp of func_8000345C_405C

+++ include/functions.us.h
@@ -562,7 +562,7 @@ void func_80132740_1416F0(Unk80160080 *arg0);
 void func_801330E4_142094(Unk80160080 *arg0);
 void func_800B33BC_C236C(s32 arg0);
 void func_80135CB8_144C68(void);
-void func_8000345C_405C(s32 arg0);
+void func_8000345C_405C(u16 arg0);
 void func_801358E8_144898(Unk80160080 *arg0, s16 arg1, s16 arg2);
 void func_801333D0_142380(void);
 void func_801336CC_14267C(Unk80160080 *arg0, Vec3f *arg1);
